Get latest data from Washington Post Github Repo

Checks if new data is available from the Washington Post Github repo and returns the most current dataset.

Usage

latest_data()

Value

tibble

Examples

latest_data()
#> [1] "No update available. Returning existing package dataset."
#> # A tibble: 338 × 50
#>      uid nces_s…¹ schoo…² nces_…³ distr…⁴ date       schoo…⁵  year time  day_o…⁶
#>    <dbl> <chr>    <chr>   <chr>   <chr>   <date>     <chr>   <dbl> <tim> <chr>  
#>  1     1 0804800… Columb… 0804800 Jeffer… 1999-04-20 1998-1…  1999 11:19 Tuesday
#>  2     2 2200540… Scotla… 2200540 East B… 1999-04-22 1998-1…  1999 12:30 Thursd…
#>  3     3 1304410… Herita… 1304410 Rockda… 1999-05-20 1998-1…  1999 08:03 Thursd…
#>  4     4 4218990… John B… 4218990 Philad… 1999-10-04 1999-2…  1999 10:00 Monday 
#>  5     5 2502790… Dorche… 2502790 Boston  1999-11-03 1999-2…  1999 07:40 Wednes…
#>  6     6 3500690… Deming… 3500690 Deming… 1999-11-19 1999-2…  1999 12:45 Friday 
#>  7     7 4011970… Fort G… 4011970 Fort G… 1999-12-06 1999-2…  1999 07:45 Monday 
#>  8     8 1201530… Ridgew… 1201530 Pasco   2000-01-19 1999-2…  2000 15:00 Wednes…
#>  9     9 0624600… Alicia… 0624600 Merced… 2000-01-26 1999-2…  2000 08:34 Wednes…
#> 10    10 1709930… Duke E… 1709930 City o… 2000-02-14 1999-2…  2000    NA Monday 
#> # … with 328 more rows, 40 more variables: city <chr>, state <chr>,
#> #   school_type <chr>, enrollment <dbl>, killed <dbl>, injured <dbl>,
#> #   casualties <dbl>, shooting_type <chr>, age_shooter1 <dbl>,
#> #   gender_shooter1 <chr>, race_ethnicity_shooter1 <chr>,
#> #   shooter_relationship1 <chr>, shooter_deceased1 <dbl>,
#> #   deceased_notes1 <chr>, age_shooter2 <dbl>, gender_shooter2 <chr>,
#> #   race_ethnicity_shooter2 <chr>, shooter_relationship2 <chr>, …